+++ js/data.js
@@ -1,0 +1,37 @@
+const DATA = {
+  COMMENT_LIST : [
+    'Всё отлично!',
+    'В целом всё неплохо. Но не всё.',
+    'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+    'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+    'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+    'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+  ],
+  NAMES_LIST : [
+    'Артем',
+    'Данил',
+    'Света',
+    'Алексей',
+    'Павел',
+    'Екатерина',
+    'Лида',
+    'Андрей',
+    'Пудж'
+  ],
+  DESCRIPTION_LIST : [
+    'Хорошая фотка!',
+    'Насыщенность отличная',
+    'Лучшее место!',
+    'Это просто вау!',
+    'Я старался',
+    'Вполне неплохо получилось'
+  ],
+  PHOTOS_COUNT : 25,
+  COMMENT_MESSAGE_COUNT : 2,
+  COMMENTS_COUNT : 30,
+  AVATARS_COUNT : 6,
+  MIN_LIKES_COUNT : 15,
+  MAX_LIKES_COUNT : 200
+};
+
+export { DATA };
